# [core] Replace expressions RTTI with enums + static cast