Replacing all `dynamic_cast` usages in expressions code with a strategy enumerator plus `static_cast` obtained benchmark results up to 27% faster:

Benchmark | Time | CPU | Time Old | Time New | CPU Old | CPU New
-- | -- | -- | -- | -- | -- | --
Parse_CameraFunction/1 | -22.47% | -22.41% | 4343 | 3367 | 4369 | 3390
Parse_CameraFunction/2 | -21.24% | -21.08% | 4959 | 3906 | 4979 | 3929
Parse_CameraFunction/4 | -22.06% | -21.72% | 6510 | 5074 | 6503 | 5090
Parse_CameraFunction/6 | -24.25% | -23.88% | 7880 | 5969 | 7873 | 5993
Parse_CameraFunction/8 | -23.66% | -23.45% | 9230 | 7046 | 9224 | 7062
Parse_CameraFunction/10 | -27.79% | -27.50% | 10453 | 7549 | 10448 | 7574
Parse_CameraFunction/12 | -27.85% | -27.63% | 11652 | 8408 | 11649 | 8431
Evaluate_CameraFunction/1 | 2.43% | 2.73% | 56 | 58 | 56 | 58
Evaluate_CameraFunction/2 | -0.26% | 0.00% | 167 | 167 | 167 | 167
Evaluate_CameraFunction/4 | 0.25% | 0.52% | 222 | 223 | 222 | 223
Evaluate_CameraFunction/6 | 0.31% | 0.51% | 243 | 244 | 243 | 244
Evaluate_CameraFunction/8 | 0.57% | 0.76% | 254 | 255 | 253 | 255
Evaluate_CameraFunction/10 | 1.57% | 1.74% | 219 | 223 | 219 | 223
Evaluate_CameraFunction/12 | 0.31% | 0.47% | 181 | 182 | 181 | 182
Parse_CompositeFunction/1 | -24.82% | -24.71% | 7668 | 5765 | 7710 | 5804
Parse_CompositeFunction/2 | -23.64% | -23.48% | 12984 | 9914 | 13018 | 9962
Parse_CompositeFunction/4 | -21.64% | -21.58% | 28076 | 21999 | 28111 | 22045
Parse_CompositeFunction/6 | -19.38% | -19.34% | 51737 | 41709 | 51759 | 41751
Parse_CompositeFunction/8 | -17.10% | -17.04% | 82667 | 68530 | 82709 | 68616
Parse_CompositeFunction/10 | -17.94% | -17.89% | 121898 | 100024 | 121928 | 100110
Parse_CompositeFunction/12 | -16.89% | -16.84% | 167129 | 138899 | 167152 | 139009
Evaluate_CompositeFunction/1 | -4.10% | -4.03% | 415 | 398 | 415 | 398
Evaluate_CompositeFunction/2 | -4.02% | -3.95% | 684 | 657 | 684 | 657
Evaluate_CompositeFunction/4 | -3.35% | -3.28% | 843 | 815 | 842 | 815
Evaluate_CompositeFunction/6 | -4.60% | -4.51% | 932 | 889 | 931 | 889
Evaluate_CompositeFunction/8 | -3.20% | -3.13% | 956 | 925 | 955 | 925
Evaluate_CompositeFunction/10 | -3.44% | -3.36% | 964 | 931 | 963 | 931
Evaluate_CompositeFunction/12 | -3.75% | -3.66% | 1013 | 975 | 1012 | 975
Parse_SourceFunction/1 | -10.89% | -10.88% | 4736 | 4220 | 4770 | 4251
Parse_SourceFunction/2 | -5.98% | -5.94% | 5232 | 4919 | 5263 | 4950
Parse_SourceFunction/4 | -8.71% | -8.70% | 6355 | 5802 | 6383 | 5828
Parse_SourceFunction/6 | -9.20% | -9.21% | 7508 | 6817 | 7535 | 6841
Parse_SourceFunction/8 | -8.12% | -8.20% | 8414 | 7731 | 8451 | 7758
Parse_SourceFunction/10 | -8.82% | -8.98% | 9380 | 8552 | 9428 | 8581
Parse_SourceFunction/12 | -8.63% | -8.73% | 10438 | 9538 | 10475 | 9561
Evaluate_SourceFunction/1 | -3.80% | -3.88% | 374 | 360 | 374 | 360
Evaluate_SourceFunction/2 | -3.36% | -3.37% | 478 | 462 | 479 | 462
Evaluate_SourceFunction/4 | -2.63% | -2.68% | 520 | 506 | 520 | 506
Evaluate_SourceFunction/6 | -0.95% | -1.00% | 542 | 537 | 543 | 537
Evaluate_SourceFunction/8 | -2.84% | -2.86% | 552 | 536 | 552 | 536
Evaluate_SourceFunction/10 | -1.94% | -1.98% | 548 | 537 | 548 | 537
Evaluate_SourceFunction/12 | -2.66% | -2.71% | 560 | 545 | 561 | 545

/cc @mapbox/gl-core @mapbox/embedded