@anandthakker 2018-07-23
I think the differences in the `Evaluate` benchmarks must be due to speeding up `expression::isFeatureConstant()` and `expression::isZoomConstant()`.  Given that, should we consider simply storing those two as properties on each expression?