@brunoabinader 2018-07-23
I took the name out of the [strategy design pattern](https://en.m.wikipedia.org/wiki/Strategy_pattern), but I'm also fine with Kind (that was also one of the options I had in mind!)