@jfirebaugh 2018-07-23
Can you trigger a fresh build (e.g. rebase on master) so we can get the size diff (currently size diff only gets reported if the PR was already open; this won't be necessary after https://github.com/mapbox/mapbox-gl-native/pull/12446).