@brunoabinader 2018-07-23
Size checking results:

| Binary         | Size change                |
|-|-|
| Android AAR    | -17.5 kB -0.173% (10.1 MB) |
| Android arm-v7 | -12.3 kB -0.264% (4.64 MB) |
| Android arm-v8 | -65.5 kB -0.951% (6.83 MB) |
| Android x86    | -16.4 kB -0.240% (6.82 MB) |
| Android x86_64 | -16.4 kB -0.238% (6.88 MB) |
| iOS Dynamic    | -16.5 kB -0.088% (18.6 MB) |
| iOS arm64      |     72 B -0.002% (4.68 MB) |
| iOS armv7      |    -80 B -0.002% (4.33 MB) |
| iOS x86_64     | -4.21 kB -0.086% (4.86 MB) |